ai-volunteer-mvp/prediction-service/requirements.txt
functions-framework==3.4.0
prophet==1.1.5
pandas==2.1.4
pytest==7.4.3
psycopg2-binary==2.9.9
